task18/Bot_Advanced/requirements.txt
openai == 1.59.9
langchain == 0.3.15
langchain-community == 0.3.15
langchain-openai == 0.3.1
aiogram == 3.17.0
pandas == 2.2.3
nest_asyncio == 1.6.0
asyncpg == 0.30.0
aiocron == 1.8
python-dotenv == 1.0.1
requests == 2.32.3
faiss-cpu==1.9.0.post1
pandas==2.2.3
openpyxl==3.1.5
